+++ modules/lucene-monitor-helpers/java/org/apache/lucene/store/MMapDirectoryReplacementJDK.java
@@ -1,0 +1,7 @@
+package org.apache.lucene.store;
+
+/**
+ * The class to override quarkus-lucene class
+ */
+final class MMapDirectoryReplacementJDK {
+}