+++ CHANGELOG.md
@@ -9,6 +9,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - Upgrade to Lucene 9.3.0
 - Upgrade GraalVM to 22.2.0
 - Option for Lucene Monitor queries index disk storage under option `:queries-index-dir`
+- Update `lucene-text-analysis` to 1.0.17
 
 ### Changed