+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - Upgrade GraalVM to 22.2.0
 - Option for Lucene Monitor queries index disk storage under option `:queries-index-dir`
 - Update `lucene-text-analysis` to 1.0.17
+- Handle `NO_COLOR` environment variable
 
 ### Changed